Move minimalSimplify to MiscTools
